pyproject.toml
[project]
name = "fitness-coach"
version = "0.1.0"
description = "Personal AI fitness coach with Strava integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "ollama>=0.4.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "fastapi>=0.133.0",
    "uvicorn[standard]>=0.40.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.21",
    "anthropic>=0.83.0",
    "google-genai>=1.64.0",
]

[project.scripts]
fitness-coach = "fitness_coach.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fitness_coach"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
